Split resolver expected-local pattern tests
